Added default 'jsconfig.json' file
A jsconfig just defines a JavaScript project for VS Code. The JavaScript experience is improved when you have a jsconfig.json file in your workspace that defines the project context. By default the JavaScript language service will analyze and provide IntelliSense for all files in your JavaScript project. You will want to specify which files to exclude or include in order to provide the proper IntelliSense.