NPM. Install packages to use `sass/scss` in and out of components
Description:
- Installs the packages needed to work with `SASS` and share stylesheet resources in a `Nuxt.js` project.
These packages are for a `Node.js` project that uses `Nuxt.js`. as development dependencies, so they will be included in the `devDependencies` file of your project's `package.json` file.
Here is a description of each package:
`sass`: is a stylesheet preprocessor that allows you to write CSS styles in a more advanced and organized way.
`sass-loader`: is a webpack module that allows developers to use SASS in their Nuxt.js projects.
`@nuxtjs/style-resources`: is a Nuxt.js module that allows developers to share variables, mixins and other common resources between different stylesheet files.

Steps:
- Install the npm packages `sass`, `sass-loader` and `@nuxt/style-resources`.
Use the command `npm install sass sass-loader @nuxtjs/style-resources --save-dev` for this.