CONFIG. Add nuxt configuration to use `sass/scss` outside components
Description:
- Adds configuration for the `@nuxtjs/style-resources` module in a Nuxt.js based project.
The `@nuxtjs/style-resources` module allows sharing variables, mixins and other stylesheet resources in different stylesheet files in a project.
The bracketed object provides additional options for manipulating the module's behavior.
In this case, you are specifying an SCSS file (a variant of SASS) as a shared resource, with the file `styles.scss` in the `@/assets/scss/styles.scss` directory.

Steps:
- Add `@nuxtjs/style-resources` in the `module` object of the nuxt configuration.
- Pass additional options as an object to the `@nuxtjs/style-resources` module configuration.
- Add the path to the global scss styles for the whole project.
- Create the `scss` folder inside the `assets` folder.
- Create the `styles.scss` file in the `scss` folder.