CONFIG. Add configuration to use svg sprites in nuxt
Steps:
- Add to the nuxt configuration in the `buildModules` object the `@nuxtjs/svg-sprite` module to use it.
The `@nuxtjs/svg-sprite` module allows to generate an SVG sprite (a file containing several vector graphics) from SVG files in a specified directory.
- In addition to the module name you are provided with an object with additional options to manipulate the module's behavior.
In this case, you are specifying the `input` property as `~/assets/images/icons/`, to indicate the input directory for the SVG files that will be used to generate the sprite.