REFACTOR. Extract code to the new `LogoList` component
Steps:
- Create the `LogoList.vue` file.
- Import and use the `LogoList` component inside the `about` page.
- Move the html structure and css styles from the `about` page to the `LogoList` component.
- In the `LogoList` component:
Update the html structure using the `ul` and `li` tags, with the `svg-icon` component inside the `li` tag.
Create the `LogoList` prop as `required` and `Array`, to print the data with a loop using the `v-for` directive.
Overwrites default styles applied by the browser to `ul` and `li`.
Bad use of BEM, you are using the `logo` component as an element and the `logo__list` element as a component, when it is the other way around. Change the name of the `logo` class to `logo-list`.
And it uses the `icon` class of the `svg-icon` component to size the svg instead of creating a new class for it.
- In the `about` page, create the `logoList` data with the names of the svg files, which are used to call them from the sprite, and pass them as prop to the `logoList` component.