BUGFIX. Get error when execute the `npm run test` command.
Error:
- Cannot find module 'babel-core'

Solution:
- Jest still uses the older version of Babel (6.x) and does not recognize the newer version of `@babel/core` (v7+).
To work around this problem, you can install an older version of `@babel/core` that is compatible with Jest. In particular, you will need to install `@babel-core@^7.0.0-bridge.0`, which is a bridge package to allow Jest to use the latest version of `@babel/core`.
You can do this by running the following command in your terminal:
`npm install --save-dev babel-core@^7.0.0-bridge.0`