BUGFIX. Found error in the `dependabot.yml` file
Error:
- The error you encountered in your `.github/dependabot.yml` file is related to incorrect formatting and usage of the `commit-message` section.

Solution:
- To fix the issue and conform with Dependabot's specification, you need to make the following adjustments:
Remove the `include` property from the `commit-message` section as it is not allowed. Dependabot only supports a fixed commit message format and doesn't provide an option to include/exclude specific messages.
Ensure that the `commit-message` section only contains the `prefix` and `prefix-development` properties. These properties define the prefixes to be used for different types of commits.

References:
- https://github.com/beatrizsmerino/nuxt-course/issues/389