fix(project): `textarea` and `select` fields do not print their data
Reason:
- `FormField` is a dynamic component that can be an html tag `input` or `textarea`.
Both display their data differently, the input needs to use the `value` attribute and the `textarea` needs to put the value between its braces.
As the code is now I cannot correct the `textarea` without breaking the `input`.

Solution:
- In the `PostForm` component the `fieldValue` prop is used instead of `v-model` to send the data to the `FormField` component.
- In the `FormField` component `v-model` is used instead of `:value` and `@input` to receive and update data.
It is necessary to use `v-model` directly in the `textarea`, it is not allowed to use interpolation between its tags to display the data.
The `updateFieldValue` computed property uses the `get` functions to get the value of `fieldValue` and `set` to return the new value to the parent `PostForm` component.
- Finally, it was necessary to use `v-if` conditionals instead of a dynamic component, otherwise it would not be reactive to changes.
- The same procedure has been followed with the `FormSelect` component.