fix(project): get error about `crypto.randomUUID()` when reload `admin/post/create` page
Error:
- ReferenceError: crypto is not defined
"id": crypto.randomUUID(),

Solution:
- Create the `getId()` method and execute it on the `mounted` hook.
This way, the UUID will be generated only after the component has been mounted in the browser, which never happens on the server.

Reason:
- The `mounted()` method in Vue.js is a lifecycle hook that is called after an instance of the component has been mounted in the DOM. Unlike other lifecycle hooks such as `created` or `serverPrefetch` which are executed on both the server and the client in the context of a Nuxt.js application with server-side rendering (SSR), the `mounted` hook is executed only on the client side.
This means that when the code inside the `mounted()` is executed, the server-side rendering process has already finished and the application is now running in the user's browser. Since `crypto` and its `randomUUID()` method are part of the browser's Web API, they are only available in this context and not on the server. So when you put the code to generate a UUID inside `mounted()`, you are ensuring that it runs in an environment where the `crypto` object is defined and accessible, thus avoiding the `crypto is not defined` error that occurs if you try to access it on the server.
In short, `mounted()` works for this case because it ensures that code that depends on browser APIs is only executed when the component has been mounted on the client, where those APIs are available.