fix(project): post `date` is not updated when editing a post
Error:
- When you define an object in the `data` option of a Vue component, what you are doing is setting the initial state of that component.
The code inside the data function is executed once when the component is initializing.
This means that `getDate()` is called only once during component initialization, and the value returned by `getDate()` is assigned to `form.date`. This value will remain the same unless it is explicitly updated somewhere in the component's lifecycle or in response to some user action.

Solution:
- In order for `form.date` to be updated each time the post is saved, you must call `getDate()` at the time you want the date to be updated, such as just before making a post or save request.
In order for the date to be updated each time you edit a post, you must make sure that the date value is set just before sending the updated data to `Firebase`.
This means that you will need to update the date inside the method just before making the `axios` call.