fix(project): missing post `id` in `PostPreview` after creating a post
Error:
- After creating a post, the `id` of the `post` is set correctly in `firebase` but appears as `undefined` in the vuex `state`.
The problem is in the `fetchCreatePost` vuex `action`. When you create the post, you get an `id` that `firebase` automatically generates. However, in `setCreatePost` vuex `mutation`, when you use `resultUpdated.data.name` as `id`, it is incorrect as it does not return a `name` property.

Solution:
- Instead, you should directly use the `firebaseId` you got from the initial `POST`.
It is important to remember that if you use `vuexContext.commit("setCreatePost", resultUpdated);`, you are trying to put the entire `axios` response object into your vuex `state`, which is not appropriate. Instead, you only want the post data and the correct `id`.