feat(config): uses the `plugins` section of `nuxt.config.js` to create `filters`
Steps:
- Store the date of the post without formatting.
Use `new Date()` instead of the mixin function `getDate()`.
- Update the mixin function `getDate()`.
Rename `getDate()` to `getDateFormatted`.
Add to `getDate()` the `date` parameter, this will be the date of the saved post.
- Create the `dateFormatted` filter in the `plugins/date-filters.js` file.
Import the mixin function `getDateFormatted` into the `plugins/date-filters.js` file.
Define a new vue filter with the name `dateFormatted` and with the `getDateFormatted` function to execute.
Add the `plugins/date-filters.js` file to the `plugins` section of the `nuxt.config.js` file.
- Use the `dateFormatted` filter in the `PostPreview` and `PostDetail` components.
Now instead of using the `getDate()` mixin function, the `dateFormatted` filter will be used.

References:
- https://v2.vuejs.org/v2/guide/filters
- https://vinodugale219.medium.com/add-date-filter-in-nuxtjs-7dead1373d7c