feat(config): use `modules` section of `nuxt.config.js` to configure `axios`
Steps:
- To use `axios` as a `module` you need to:
Install the `@nuxtjs/axios` package.
Add `@nuxtjs/axios` to the `modules` section.
- The `@nuxtjs/axios` options can be integrated inside or outside the `modules` section.
Each module can be added inside an array with its name and the options to configure it.
Or you can put only the module name inside the `modules` section and outside it create an object with the module options.
- Update the `baseURL` property of the `@nuxtjs/axios` module.
Charge the ENV variable `BASE_URL` or the `firebase` path to the project database.
- Add the `credentials` property of the `@nuxtjs/axios` module.
Disable the `credentials` property with the value `false`, since it does not need authentication headers to get the data from the firebase database.
- Update the use of `axios` in the `store/index.js` file.
It is no longer necessary to import `axios` since it is integrated in the project.
Use `context.app.$axios.$get` or `this.$axios.$get` instead `axios.get`.
Get the data directly, instead of using `response.data` or `result.data`.
Now it is no longer necessary to use the `process.env.baseURL` in the path to request the data, since `axios` loads directly that path that we have configured in the module with the `baseURL` property.

References:
- https://nuxt.com/modules
- https://v2.nuxt.com/docs/directory-structure/modules/
- https://v2.nuxt.com/docs/directory-structure/modules#write-your-own-module
- https://axios.nuxtjs.org/setup#configure
- https://axios.nuxtjs.org/options#baseurl
- https://axios.nuxtjs.org/usage#store-actions
- https://axios.nuxtjs.org/options#credentials