feat(project): save in the `localStorage` the `tokenId` and the `tokenExpire`
Steps:
- In the `fetchCreateAuthUser` vuex `action` set to the `localStorage` the `authTokenId` and the `authTokenExpire`.
Calculates the current time plus the expiration time to obtain the time in milliseconds to expiration.
- Create the `fetchReadAuthUser` vuex `action` get from the `localStorage` the `authTokenId` and the `authTokenExpire`.
If the current time is greater than the expiration time or the token does not exist, do nothing.
If not, run `fetchDeleteAuthUser` vuex `action` calculating the time the token expires minus the current time.
And also if not, run `fetchReadAuthAuthUser` vuex `action` to save the `authTokenId` in the `localStorage`.
- Create `auth-check` nuxt `middleware` to run the `fetchReadAuthUser` vuex `action` only when is the client-side because the `localStorage` only exist there.
- Add `auth-check` nuxt `middleware` to all `admin` pages except the `admin/auth` page.
- Add `console.log` to `auth-redirect` and `auth-check` nuxt `middleware` to check when they are executed.