fix(project): when refreshing the page the list of posts disappears
Error:
- This fixes the problem that when refreshing the page the list of posts disappears.

Solution:
- Use `Cookies` for the `server-side` and `localStorage` for the `client-side`.

Steps:
- It is necessary to install and import `js-cookie` package in the `store/index.js` file.
- In the `fetchCreateAuthUser` vuex `action`, when save the `authTokenId` and `authTokenExpire` in the `localStorage` also save as `Cookie`.
- The `fetchReadAuthUser` vuex `action` is executed in the `auth-check` vuex `middleware`.
If we are on the `server-side` we pass the `request` but if we are on the `client-side` we pass `null`.
- In the `fetchReadAuthUser` vuex `action`, we pass a second param, `request`, to check if it is using the `server-side` or `client-side`.
If `request` is `null` we use `localStorage` to get the `authTokenId` and `authTokenExpire`. But if the `request` is not `null` we use `Cookies` to get the `authTokenId` and `authTokenExpire`.
Convert the `tokenId` and `tokenExpire` from `const` to `let`, in order to override their value depending on the `request` parameter.
Gets the `headers.cookie` from `request` if saved, to get the `authTokenId` and `authTokenExpire` from the `Cookies`.