feat(project): create a `serverMiddleware` in the `nuxt.config.js`
Steps:
- It must have the `express` and `body-parser` packages installed to work.
Import the `express` package in the `api/index.js` file.
Import the `body-parser` package in the `nuxt.config.js` file.
- Create the `api/index.js` file to set up a router with an API path, which handles POST requests in the `/track-data` path and then exports that API path so it can be used by other files in the application.
Create a router using `express.Router()` to define the API routes.
Create an Express application (`app`) using `express()`.
Add middleware to the router to configure the `request` and `response` objects for each incoming request to customize the behavior of these objects during request handling.
Define a `POST` route in the router for the `/track-data` route. When a `POST` request is received on this path, the data stored in the request body (`request.body.data`) is printed to the console and a JSON response is sent with a 200 status code and a success message.
Define an `apiRoute` object containing the route and router. This data structure will probably be used to export the API routes and make them accessible from other files.
Export the `apiRoute` object as a default value, which means that when this file is imported into another module, the exported value will be `apiRoute`.
- Add the `serverMiddleware` property to the `nuxt.config.js` file.
Add to the `serverMiddleware` property the `bodyParser.json()`, to parse JSON data in incoming requests, and the `~/api`, to define additional server paths or API endpoints in the application.

References:
- https://v2.nuxt.com/docs/configuration-glossary/configuration-servermiddleware/#custom-api-endpoint