fix(github-actions): `babel` error when run `lint` command
Error:
- Found error in `env-package.js`:
Parse errors in imported module './mixins/repo-mixins.js': The `assert` keyword in import attributes is deprecated and it has been replaced by the `with` keyword. You can enable the `deprecatedAssertSyntax: true` option in the import attributes plugin to suppress this error.
- Found error in `repo-mixins.js`:
Parsing error: The `assert` keyword in import attributes is deprecated and it has been replaced by the `with` keyword. You can enable the `deprecatedAssertSyntax: true` option in the import attributes plugin to suppress this error.

Reason:
- The `importAssertions` syntax (`assert { type: "json" }`) that was previously used is being replaced by the `importAttributes` syntax (`with { type: "json"}`). These two syntaxes are not compatible and cannot coexist in the same `babel` configuration.

Solution:
- Add the `deprecatedAssertSyntax: true` option to `@babel/plugin-syntax-import-attributes` package from the configuration file: `.babelrc`.

References:
- https://babeljs.io/docs/babel-plugin-syntax-import-attributes
- https://babeljs.io/docs/babel-plugin-proposal-import-attributes-to-assertions