fix(stylelint): `scss` is not formatted on save file when run `dev` command
Error:
- DeprecationWarning: The CommonJS Node.js API is deprecated.
See https://stylelint.io/migration-guide/to-16
(Use node --trace-deprecation ... to show where the warning was created)

Reason:
- When upgrading the `stylelint` package to version 16, the `lint` command still worked fine, as it runs `stylelint` independently to format all files manually.
However, when running the `dev` command, the problem occurs because it uses the `@nuxtjs/stylelint-module` to integrate `stylelint` into the `nuxt` development flow to format the modified files each time they are saved.
The `@nuxtjs/stylelint-module@4` package does not work because it is not compatible with `stylelint@16`, it needs to be upgraded to `@nuxtjs/stylelint-module@5`.

Solution:
- To solve this problem, it is necessary to upgrade to `@nuxtjs/stylelint-module@5` and consequently to `nuxt@3` and `vue@3` as this package depends on these versions.
But as the aim of this project is to practice and learn `nuxt@2` and `vue@2`, we can't do that.
Instead, we need to downgrade to `stylelint@15`.

Reverts part of the commit: `127941dc4ca0f81a6ed00e6357d3db449db0213d`